Use new Peregrine API to create store
